requirements.txt
beautifulsoup4>=4.12
lxml>=5.0
pandas>=2.0
plotly>=5.18
streamlit>=1.30
pytz>=2024.1
python-telegram-bot>=21.0
python-dotenv>=1.0
pyarrow>=14.0
